requirements.txt
PyPDF2
tabula-py
tabula-py[jpype]
pandas
